Re-work file saving
When building under GCC 10.2 on Fedora 32 the following compiler warning
was being produced

reordinator.c: In function ‘cb_save_file’:
reordinator.c:200:43: warning: ‘%d’ directive output may be truncated writing between 1 and 11 bytes into a region of size between 0 and 4095 [-Wformat-truncation=]
  200 |  snprintf(tmp_file, sizeof(tmp_file), "%s.%d.tmp", loaded_file,
      |                                           ^~
reordinator.c:200:2: note: ‘snprintf’ output between 7 and 4112 bytes into a destination of size 4096
  200 |  snprintf(tmp_file, sizeof(tmp_file), "%s.%d.tmp", loaded_file,
      |  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  201 |    getpid());
      |    ~~~~~~~~~

Which is quite correct, loaded_file could be upto 4096 bytes and
tmp_file is that size so we could truncate adding the .PID.tmp suffix.

There are numerous ways to deal with this, in the end I decided to
re-work the file saving by splitting loaded_file into its directory and
filename components (just in cb_save_file()) and using the openat(2) &
renameat(2) system calls.

Truncation of the tmp_file name could still happen if the original
filename was around 249 characters long but this should still get
caught when we attempt to open it and inform the user.

We now create this temporary file as hidden (. prefixed).

Also include the right limits.h for NAME_MAX etc...

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>