Fix open(2) flags in cb_save_file()
This open() is used to get a file descriptor for the directory
containing the file being being edited for use in openat(2)

I had the flags for this as O_DIRECTORY, however O_DIRECTORY is not
specifically telling open() you're opening a directory, it works without
that, but that what you're opening _should_ be a directory and open()
will return an error if that's not the case.

In this case we have used dirname(3) to get the directory, so change
O_DIRECTORY to O_RDONLY (that was being added (by glibc I guess)).
O_RDONLY seems to work OK, regardless if we're opening a file reading or
writing in the subsequent openat().

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>